Intersersection Quality of life
Changed data type of cycleTime to int from ulong
When the cycleTime leaves bounds it is reset to a random value not just
100

+++ TrafficSim/StartingValues.cs
@@ -7,7 +7,7 @@ namespace TrafficSim
         public byte MapHeight;
         public ulong Budget;
         public ulong Population;
-        public ulong TrafficLightCycleTimeDefault;
+        public int TrafficLightCycleTimeDefault;
         public string StoragePath;
         public string StoragePathMap;
     }
